# docs/open/openapi.yml
openapi: "3.0.3"
info:
  title: "HTTP API"
  version: "6.0"
servers:
  - url: "http://localhost:5000/api/v6"
security:
  - jwt: []
  - ApiKeyAuth: []
...
paths:
...
  /currency:
    get:
      tags:
        - currency
      summary: List currencies
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CurrencyResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - currency
      summary: Create a new currency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/CurrencyPost"
      responses:
        "201":
          description: Currency created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /customer:
    get:
      tags:
        - customer
      summary: List customers
      parameters:
        - name: customer_type
          in: query
          schema:
            type: string
            enum: [CUSTOMER_COMPANY, CUSTOMER_PRIVATE, CUSTOMER_OTHER, CUSTOMER_OWN]
            description: Customer type enumeration
        - name: customer_name
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - customer
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/CustomerPost"
      responses:
        "201":
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /item:
    get:
      tags:
        - item
      summary: List items
      parameters:
        - name: trans_code  
          in: query
          schema:
            type: string
        - name: product_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'    
    post:
      tags:
        - item
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/ItemPost"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /link:
    get:
      tags:
        - link
      summary: List links
      parameters:
        - name: link_type_1  
          in: query
          schema:
            type: string
            enum: [LINK_CUSTOMER, LINK_EMPLOYEE, LINK_ITEM, LINK_MOVEMENT, LINK_PAYMENT, LINK_PLACE, LINK_PRODUCT, LINK_PROJECT, LINK_TOOL, LINK_TRANS]
        - name: link_code_1
          in: query
          schema:
            type: string
        - name: link_type_2
          in: query
          schema:
            type: string
            enum: [LINK_CUSTOMER, LINK_EMPLOYEE, LINK_ITEM, LINK_MOVEMENT, LINK_PAYMENT, LINK_PLACE, LINK_PRODUCT, LINK_PROJECT, LINK_TOOL, LINK_TRANS]
        - name: link_code_2
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - link
      summary: Create a new link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/LinkPost"
      responses:
        "201":
          description: Link created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MetaTags:
      type: array
      items:
        type: string
      description: Additional tags for the data
      example: ["TAG1", "TAG2"]
    DataMap:
      type: object
      additionalProperties: true
      description: Flexible key-value map for additional metadata. The value is any json type.
      example:
        field_name: "Field value"
    Address:
      type: object
      properties:
        country:
          type: string
        state:
          type: string
        zip_code:
          type: string
        city:
          type: string
        street:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
    Contact:
      type: object
      properties:
        first_name:
          type: string
        surname:
          type: string
        status:
          type: string
        phone:
          type: string
        mobile:
          type: string
        email:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
    Event:
      type: object
      properties:
        uid:
          type: string
        subject:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        place:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
...
    Currency:
      type: object
      properties:
        currency_meta:
          type: object
          properties:
            description:
              type: string
              description: The name of the currency
            digit:
              type: integer
              format: int64
              description: Number of decimal places for recording and rounding
            cash_round:
              type: integer
              format: int64
              description: Rounding value for cash transactions
            tags:
              $ref: "#/components/schemas/MetaTags"
        currency_map:
          $ref: "#/components/schemas/DataMap"
    CurrencyPost:
      type: object
      required:
        - code
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. The ISO 4217 code of the currency. The value is always mandatory
              example: "EUR"
        - $ref: "#/components/schemas/Currency"
    CurrencyResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
              example: 12345
        - $ref: "#/components/schemas/CurrencyPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of currency creation
              example: "2023-06-15T10:30:15Z"
    Customer:
      type: object
      properties:
        customer_type:
          type: string
          enum: [CUSTOMER_COMPANY, CUSTOMER_PRIVATE, CUSTOMER_OTHER, CUSTOMER_OWN]
          example: "CUSTOMER_COMPANY"
          description: Customer type enumeration
        customer_name:
          type: string
          description: Full name of the customer
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        customer_meta:
          type: object
          properties:
            tax_number:
              type: string
            account:
              type: string
            tax_free:
              type: boolean
              description: Tax-free status
            terms:
              type: integer
              format: int64
              description: Payment per.
            credit_limit:
              type: number
              format: double
              description: Customer's credit limit. Data is used by financial reports.
            discount:
              type: number
              format: double
              description: If new product line is added (offer, order, invoice etc.) all products will receive the discount percentage specified in this field.
            notes:
              type: string
            inactive:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        customer_map:
          $ref: "#/components/schemas/DataMap"
    CustomerPost:
      type: object
      required:
        - customer_type
        - customer_name
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "CUS1731101982N123"
        - $ref: "#/components/schemas/Customer"
    CustomerResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
              example: 12345
        - $ref: "#/components/schemas/CustomerPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of customer creation
              example: "2023-06-15T10:30:15Z"
...
    Item:
      type: object
      properties:
        trans_code:
          type: string
          description: Reference to trans.code
          example: "TRN1731101982N123"
        product_code:
          type: string
          description: Reference to product.code
          example: "PRD1731101982N123"
        tax_code:
          type: string
          description: Reference to Tax.code
          example: "TAX1731101982N123"
        item_meta:
          type: object
          properties:
            unit:
              type: string
            qty:
              type: number
              format: double
            fx_price:
              type: number
              format: double
            net_amount:
              type: number
              format: double
            discount:
              type: number
              format: double
            vat_amount:
              type: number
              format: double
            amount:
              type: number
              format: double
            description:
              type: string
            deposit:
              type: boolean
            own_stock:
              type: number
              format: double
            action_price:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        item_map:
          $ref: "#/components/schemas/DataMap"
    ItemPost:
      type: object
      required:
        - trans_code
        - product_code
        - tax_code
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "ITM1731101982N123"
        - $ref: "#/components/schemas/Item"
    ItemResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
              example: 12345
        - $ref: "#/components/schemas/ItemPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
              example: "2023-06-15T10:30:15Z"
    Link:
      type: object
      properties:
        link_type_1:
          type: string
          enum: [LINK_CUSTOMER, LINK_EMPLOYEE, LINK_ITEM, LINK_MOVEMENT, LINK_PAYMENT, LINK_PLACE, LINK_PRODUCT, LINK_PROJECT, LINK_TOOL, LINK_TRANS]
          description: ENUM field specifying the first link type
        link_code_1:
          type: string
          description: Reference to LinkType1.code
        link_type_2:
          type: string
          enum: [LINK_CUSTOMER, LINK_EMPLOYEE, LINK_ITEM, LINK_MOVEMENT, LINK_PAYMENT, LINK_PLACE, LINK_PRODUCT, LINK_PROJECT, LINK_TOOL, LINK_TRANS]
          description: ENUM field specifying the second link type
        link_code_2:
          type: string
          description: Reference to LinkType2.code
        link_meta:
          type: object
          properties:
            qty:
              type: number
              format: double
            rate:
              type: number
              format: double
            tags:
              $ref: "#/components/schemas/MetaTags"
        link_map:
          $ref: "#/components/schemas/DataMap"
    LinkPost:
      type: object
      required:
        - link_type_1
        - link_code_1
        - link_type_2
        - link_code_2
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "LNK1731101982N123"
        - $ref: "#/components/schemas/Link"
    LinkResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/LinkPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
...